+++ halo2_proofs/src/dev/graph/layout.rs
@@ -4,16 +4,18 @@ use plotters::{
     coord::Shift,
     prelude::{DrawingArea, DrawingAreaErrorKind, DrawingBackend},
 };
-use std::cmp;
 use std::collections::HashSet;
 use std::ops::Range;
 
 use crate::{
+    circuit::layouter::RegionColumn,
     circuit::{layouter::RegionColumn, Value},
+    dev::cost::Layout,
     plonk::{
         Advice, Any, Assigned, Assignment, Circuit, Column, ConstraintSystem, DynamicTable,
         DynamicTableInfo, Error, Fixed, FloorPlanner, Instance, Selector,
     },
+    plonk::{Any, Circuit, Column, ConstraintSystem, FloorPlanner},
 };
 
 /// Graphical renderer for circuit layouts.