+++ halo2_proofs/CHANGELOG.md
@@ -6,6 +6,10 @@ and this project adheres to Rust's notion of
 [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- `halo2_proofs::circuit::layouter`:
+  - `RegionLayouter::instance_value` method added to provide access to
+    instance values within a region.
 
 ## [0.2.0] - 2022-06-23
 ### Added